split out TabPanel component
